requirements.txt
# Core
django==2.0
djangorestframework==3.7.3
arrow==0.12.0
django-phonenumber-field==1.3.0
django-model-utils==3.0.0
django-environ-0.4.4
psycopg2==2.7.3.2

# Testing
factory-boy==2.9.2
pytest==3.3.0
flake8==3.5.0
pytest-django==3.1.2
pytest-sugar==0.9.0
